% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{ddirmult}
\alias{ddirmult}
\title{PMF of Dirichlet-multinomial distribution}
\usage{
ddirmult(x, alpha, lg = FALSE)
}
\arguments{
\item{x}{The vector of counts.}

\item{alpha}{The vector of concentration parameters.}

\item{lg}{A logical. Should we log the density (\code{TRUE}) or not
(\code{FALSE})?}
}
\description{
PMF of Dirichlet-multinomial distribution
}
\examples{
ddirmult(c(1, 2, 3), c(1, 1, 1))
ddirmult(c(2, 2, 2), c(1, 1, 1))

}
\author{
David Gerard
}
