% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_gl}
\alias{gibbs_gl}
\title{Gibbs sampler under random mating using genotype log-likelihoods.}
\usage{
gibbs_gl(gl, alpha, B = 10000L, T = 100L, more = FALSE, lg = FALSE)
}
\arguments{
\item{gl}{The matrix of genotype log-likelihoods. The columns index the
dosages and the rows index the individuals. \code{gl[i,j]} is the
genotype log-likelihood for individual i at dosage j. It is assumed
that natural log is used.}

\item{alpha}{Vector of hyperparameters for the gamete frequencies.
Should be length (x.length() - 1) / 2 + 1.}

\item{B}{The number of sampling iterations.}

\item{T}{The number of burn-in iterations.}

\item{more}{A logical. Should we also return posterior draws (\code{TRUE})
or not (\code{FALSE}).}

\item{lg}{Should we return the log marginal likelihood (true) or not
(false).}
}
\value{
A list with some or all of the following elements
\itemize{
\item{\code{mx}: The estimate of the marginal likelihood}
}
}
\description{
Gibbs sampler under random mating using genotype log-likelihoods.
}
\examples{
gl <- matrix(-abs(rnorm(20)), ncol = 5)
alpha <- rep(1, 3)
B <- 1
T <- 1
more <- FALSE
lg <- TRUE
gibbs_gl(gl = gl, alpha = alpha, B = B, T = T, more = more, lg = lg)

}
\author{
David Gerard
}
