% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{hwebrm}
\alias{hwebrm}
\title{Bayes test for random mating with known genotypes}
\usage{
hwebrm(nvec, lg = TRUE)
}
\arguments{
\item{nvec}{A vector containing the observed genotype counts,
where \code{nvec[[i]]} is the number of individuals with genotype
\code{i-1}. This should be of length \code{ploidy+1}.}

\item{lg}{A logical. Should we return the log Bayes factor (\code{TRUE})
or the Bayes factor (\code{FALSE})?}
}
\description{
Bayes test for random mating with known genotypes
}
\examples{
set.seed(1)
ploidy <- 8

## Simulate under the null
p <- stats::runif(ploidy / 2 + 1)
p <- p / sum(p)
q <- stats::convolve(p, rev(p), type = "open")

## See BF increase
nvec <- c(stats::rmultinom(n = 1, size = 100, prob = q))
hwebrm(nvec = nvec)

nvec <- c(stats::rmultinom(n = 1, size = 1000, prob = q))
hwebrm(nvec = nvec)

nvec <- c(stats::rmultinom(n = 1, size = 10000, prob = q))
hwebrm(nvec = nvec)

## Simulate under the alternative
q <- stats::runif(ploidy + 1)
q <- q / sum(q)

## See BF decrease
nvec <- c(stats::rmultinom(n = 1, size = 100, prob = q))
hwebrm(nvec = nvec)

nvec <- c(stats::rmultinom(n = 1, size = 1000, prob = q))
hwebrm(nvec = nvec)

nvec <- c(stats::rmultinom(n = 1, size = 10000, prob = q))
hwebrm(nvec = nvec)

}
\author{
David Gerard
}
