% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{hwefit}
\alias{hwefit}
\title{Estimate double reduction parameter and test for HWE}
\usage{
hwefit(nvec, denom = c("expected", "observed"))
}
\arguments{
\item{nvec}{A vector containing the observed genotype counts,
where \code{nvec[[i]]} is the number of individuals with genotype
\code{i-1}. This should be of length \code{ploidy+1}.}

\item{denom}{What should we use for the denominator? The expected
genotype frequencies (\code{denom = "expected"}) or the
observed genotype frequencies (\code{denom = "observed"})?}
}
\value{
A list with the following elements
\describe{
\item{\code{alpha}}{The estimated double reduction parameter(s).}
\item{\code{chisq}}{The chi-square test statistic for testing
against the null of equilibrium.}
\item{\code{p.value}}{The p-value against the null of equilibrium.}
}
}
\description{
Implements a generalized method-of-moments approach to test for
Hardy-Weinberg equilibrium in the presence of double reduction. This
also provides an estimate of double reduction given that the population
is at equilibrium.
}
\examples{
## Diploid at various levels of deviation from HWE
hwefit(c(10, 20, 10))
hwefit(c(12, 16, 12))
hwefit(c(14, 12, 14))

## Tetraploid with strong deviation from HWE
nvec <- c(25, 0, 0, 0, 25)
hwefit(nvec)

## Hexaploid with exact frequencies at HWE
nvec <- round(hwefreq(p = 0.5, alpha = 0.4, ploidy = 6, niter = 100, tol = -Inf) * 50)
hwefit(nvec)

}
\author{
David Gerard
}
