% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{hwefit_genolike}
\alias{hwefit_genolike}
\title{Equilibrium and random mating estimation and testing for many loci using
posterior genotype probabilities.}
\usage{
hwefit_genolike(genoarray, nboot = 2000)
}
\arguments{
\item{genoarray}{A three-dimensional array of posterior genotype probabilties.
The first dimension indexes
the loci, the second dimension indexes the individuals, and the
third dimension indexes the genotypes. So \code{genoarray[i,j,k]}
contains the posterior probability that individual \code{j} at locus
\code{i} has genotype \code{k-1}.}

\item{nboot}{The number of bootstrap samples to apply to
\code{\link{hweboot}()}.}
}
\description{
Right now, only the bootstrap method in \code{\link{hweboot}()} is
implemented. We hope in the future to offer more options.
}
\author{
David Gerard
}
