% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe.R
\name{hwefreq}
\alias{hwefreq}
\title{Generate HWE genotype frequencies}
\usage{
hwefreq(p, alpha, ploidy, niter = 100, tol = 10^-4)
}
\arguments{
\item{p}{The allele frequency of the reference allele.}

\item{alpha}{The double reduction parameter(s). This should be a
vector of length \code{floor(ploidy/4)} where \code{alpha[i]}
is the probability of exactly \code{i} pairs of IBDR alleles
being in the gamete. Note that \code{sum(alpha)} should be less than
1, as \code{1 - sum(alpha)} is the probability of no double reduction.}

\item{ploidy}{The ploidy of the species. This should be an even positive
integer.}

\item{niter}{The maximum number of iterations to simulate.}

\item{tol}{The stopping criterion on the Chi-square divergence between
old and new genotype frequencies.}
}
\description{
Generate genotype frequencies under Hardy-Weinberg equilibrium
given the allele frequency of the reference allele (\code{p}),
the double reduction parameter (\code{alpha}), and the ploidy
of the species (\code{ploidy}).
}
\details{
If \code{alpha} is not all 0, then this function repeatedly
applies \code{\link{freqnext}()} to simulate genotype frequencies
under HWE. Otherwise, it uses \code{\link[stats]{dbinom}()}.
}
\examples{
freq1 <- hwefreq(p = 0.5, alpha = 0, ploidy = 4)
freq2 <- hwefreq(p = 0.5, alpha = 1/4, ploidy = 4)

plot(x = 0:4,
     y = freq1,
     type = "h",
     ylim = c(0, 0.4),
     xlab = "dosage",
     ylab = "Pr(dosage)")
plot(x = 0:4,
     y = freq2,
     type = "h",
     ylim = c(0, 0.4),
     xlab = "dosage",
     ylab = "Pr(dosage)")

}
\author{
David Gerard
}
