% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like.R
\name{hwelike}
\alias{hwelike}
\title{Likelihood inference for HWE}
\usage{
hwelike(nvec)
}
\arguments{
\item{nvec}{A vector containing the observed genotype counts,
where \code{nvec[[i]]} is the number of individuals with genotype
\code{i-1}. This should be of length \code{ploidy+1}.}
}
\value{
A list with the following elements:
\describe{
\item{\code{p}}{The estimated gamete genotype frequencies. \code{p[[i}
is the estimated frequency for gamete genotype \code{i-1}.}
\item{\code{chisq_hwe}}{The likelihood ratio test statistic for testing
against the null of HWE.}
\item{\code{df_hwe}}{The degrees of freedom associated with
\code{chisq_hwe}.}
\item{\code{p_hwe}}{The p-value against the null of HWE.}
}
}
\description{
Estimates gamete genotype frequencies using a maximum likelihood approach
and runs a likelihood ratio test for HWE.
}
\examples{
## Randomly generate gamete frequencies
set.seed(1)
ploidy <- 10
pvec <- stats::runif(ploidy / 2 + 1)
pvec <- pvec / sum(pvec)

## Genotype frequencies from gamete frequencies under HWE
qvec <- stats::convolve(pvec, rev(pvec), type = "open")

## Generate data
nvec <- c(stats::rmultinom(n = 1, size = 100, prob = qvec))

## Run hwelike()
hwelike(nvec = nvec)

}
\author{
David Gerard
}
