% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{hwemom}
\alias{hwemom}
\title{Estimate double reduction parameter and test for HWE}
\usage{
hwemom(nvec, obj = c("g", "pearson", "neyman"), addval = 4/length(nvec))
}
\arguments{
\item{nvec}{A vector containing the observed genotype counts,
where \code{nvec[[i]]} is the number of individuals with genotype
\code{i-1}. This should be of length \code{ploidy+1}.}

\item{obj}{What chi-square function should we minimize?
\describe{
\item{\code{pearson}}{\deqn{\sum (o-e)^2/e}}
\item{\code{g}}{\deqn{2\sum o \log(o/2)}}
\item{\code{neyman}}{\deqn{\sum (o-e)^2/o}}
}
See Berkson (1980) for a description.}

\item{addval}{How many counts should we add to each genotype before
implementing our procedures? Defaults to \code{4 / length(nvec)},
which corresponds to the "add two" rule of Agresti and Coull (1998).}
}
\value{
A list with the following elements
\describe{
\item{\code{alpha}}{The estimated double reduction parameter(s).
In diploids, this value is \code{NULL}.}
\item{\code{chisq_hwe}}{The chi-square test statistic for testing
against the null of equilibrium.}
\item{\code{df_hwe}}{The degrees of freedom associated with
\code{chisq_hwe}.}
\item{\code{p_hwe}}{The p-value against the null of equilibrium.}
\item{\code{chisq_ndr}}{The chi-square test statistic for
testing against the null of no double reduction (conditional
on equilibrium). In diploids, this value is \code{NULL}.}
\item{\code{df_ndr}}{The degrees of freedom associated with
\code{chisq_ndr}.}
\item{\code{p_ndr}}{The p-value against the null of no double
reduction (conditional on equilibrium). In diploids, this value
is \code{NULL}.}
}
}
\description{
Implements a generalized method-of-moments approach to test for
Hardy-Weinberg equilibrium in the presence of double reduction. This
also provides an estimate of double reduction given that the population
is at equilibrium.
}
\details{
For ploides greater than 4, not all gametic frequencies are possible
given rates of double reduction. That is, if the model for meiosis
from Huang et. al. (2019) is incorrect, then this is not an appropriate
test for HWE.
}
\examples{
## Diploid at various levels of deviation from HWE
hwemom(c(10, 20, 10))
hwemom(c(12, 16, 12))
hwemom(c(14, 12, 14))

## Tetraploid with strong deviation from HWE
nvec <- c(25, 0, 0, 0, 25)
hwemom(nvec)

## Hexaploid with exact frequencies at HWE
nvec <- round(hwefreq(r = 0.5,
                      alpha = 0.2,
                      ploidy = 6,
                      niter = 100,
                      tol = -Inf) * 100)
hwemom(nvec)

## Octoploid case with exact frequencies at HWE
nvec <- round(hwefreq(r = 0.5,
                      alpha = c(0.1, 0.01),
                      ploidy = 8,
                      niter = 100,
                      tol = -Inf) * 1000)
hwemom(nvec)

}
\references{
\itemize{
\item{Agresti, A., & Coull, B. A. (1998). Approximate is better than "exact" for interval estimation of binomial proportions. The American Statistician, 52(2), 119-126. \doi{10.1080/00031305.1998.10480550}}
\item{Berkson, J. (1980). Minimum chi-square, not maximum likelihood!. The Annals of Statistics, 8(3), 457-487. \doi{10.1214/aos/1176345003}}
\item{Huang, K., Wang, T., Dunn, D. W., Zhang, P., Cao, X., Liu, R., & Li, B. (2019). Genotypic frequencies at equilibrium for polysomic inheritance under double-reduction. G3: Genes, Genomes, Genetics, 9(5), 1693-1706. \doi{10.1534/g3.119.400132}}
}
}
\author{
David Gerard
}
