% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetra.R
\name{hwetetra}
\alias{hwetetra}
\title{HWE likelihood inference for tetraploids}
\usage{
hwetetra(nvec, upperdr = 1/6)
}
\arguments{
\item{nvec}{A vector containing the observed genotype counts,
where \code{nvec[[i]]} is the number of individuals with genotype
\code{i-1}. This should be of length 5 (since we have tetraploids).}

\item{upperdr}{The upper bound on the double reduction rate. Defaults
to 1/6, which is the maximum value under the complete equational
segregation model (Mather, 1935).}
}
\value{
A list with the following elements
\describe{
\item{\code{alpha}}{The estimated double reduction rate.}
\item{\code{r}}{The estimated allele frequency.}
\item{\code{q_u}}{The genotype frequencies under the unrestricted model.}
\item{\code{ll_u}}{The log-likelihood under the urestricted model.}
\item{\code{q_rm}}{The genotype frequencies under the random mating model.}
\item{\code{ll_rm}}{The log-likelihood under the random mating model.}
\item{\code{q_hw}}{The genotype frequencies under the small deviations
from HWE model.}
\item{\code{ll_hw}}{The log-likelihood under the small deviations from HWE
model.}
\item{\code{q_ndr}}{The genotype frequencies under the no double reduction
at HWE model.}
\item{\code{ll_ndr}}{The log-likelihood under the no double reduction at
HWE model.}
\item{\code{chisq_rm}}{The chi-square statistic against the null of random
mating, with an alternative of unrestricted.}
\item{\code{df_rm}}{The degrees of freedom corresponding to \code{chisq_rm}.}
\item{\code{pval_rm}}{The p-value corresponding to \code{chisq_rm}.}
\item{\code{chisq_hw}}{The chi-square statistic against the null of only
small deviations from HWE, with an alternative of random mating.}
\item{\code{df_hw}}{The degrees of freedom corresponding to \code{chisq_hw}.}
\item{\code{pval_hw}}{The p-value corresponding to \code{chisq_hw}.}
\item{\code{chisq_ndr}}{The chi-square statistic against the null of no
double reduction at HWE, with an alternative of only small deviations
from HWE.}
\item{\code{df_ndr}}{The degrees of freedom corresponding to \code{chisq_ndr}.}
\item{\code{pval_ndr}}{The p-value corresponding to \code{chisq_ndr}.}
}
}
\description{
HWE likelihood inference for tetraploids
}
\examples{
## Generate data under HWE
set.seed(1)
alpha <- 0.1
r <- 0.4
q <- hwefreq(r = r, alpha = alpha, ploidy = 4)
nvec <- c(stats::rmultinom(n = 1, size = 500, prob = q))
hwetetra(nvec = nvec)


}
\references{
\itemize{
\item{Mather, K. (1935). Reductional and equational separation of the chromosomes in bivalents and multivalents. Journal of Genetics, 30(1), 53-78.}
}
}
\author{
David Gerard
}
