% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qqpvalue}
\alias{qqpvalue}
\title{QQ-plot for p-values}
\usage{
qqpvalue(pvals, method = c("ggplot2", "base"))
}
\arguments{
\item{pvals}{A vector of p-values.}

\item{method}{Should we use base plotting or ggplot2 (if installed)?}
}
\description{
This will create a QQ-plot for p-values, comparing them to a unifrom
distribution. We make our plot on the -log10 scale. We calculate
the confidence envelope using the method of Fox (2016), pp 37--41.
}
\examples{
set.seed(1)
pvals <- runif(100)
qqpvalue(pvals, method = "base")

}
\references{
\itemize{
\item{Fox, J. (2016) \emph{Applied Regression Analysis and Generalized Linear Models}, Third Edition. Sage.}
}
}
\seealso{
\itemize{
\item{The \code{qqPlot()} function from the car package.}
}
}
\author{
David Gerard
}
