% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{rmbayesgl}
\alias{rmbayesgl}
\title{Bayes test for random mating using genotype log-likelihoods}
\usage{
rmbayesgl(
  gl,
  lg = TRUE,
  alpha = NULL,
  beta = NULL,
  nburn = 10000,
  niter = 10000
)
}
\arguments{
\item{gl}{A matrix of genotype log-likelihoods. The rows index the
individuals and the columns index the genotypes. So \code{gl[i,k]}
is the genotype log-likelihood for individual \code{i} at
dosage \code{k-1}. We assume the \emph{natural} log is used (base e).}

\item{lg}{A logical. Should we return the log Bayes factor (\code{TRUE})
or the Bayes factor (\code{FALSE})?}

\item{alpha}{The concentration hyperparameters of the gamete frequencies
under the null of random mating. Should be length ploidy/2 + 1.}

\item{beta}{The concentration hyperparameters of the genotype frequencies
under the alternative of no random mating. Should be length ploidy + 1.}

\item{nburn}{The number of iterations in the Gibbs sampler to burn-in.}

\item{niter}{The number of sampling iterations in the Gibbs sampler.}
}
\description{
Bayes test for random mating using genotype log-likelihoods
}
\examples{

\dontrun{
set.seed(1)
ploidy <- 8

## Simulate under the null ----
p <- stats::runif(ploidy / 2 + 1)
p <- p / sum(p)
q <- stats::convolve(p, rev(p), type = "open")

## See BF increase
nvec <- c(stats::rmultinom(n = 1, size = 100, prob = q))
gl <- simgl(nvec = nvec)
rmbayesgl(gl = gl)

nvec <- c(stats::rmultinom(n = 1, size = 1000, prob = q))
gl <- simgl(nvec = nvec)
rmbayesgl(gl = gl)

## Simulate under the alternative ----
q <- stats::runif(ploidy + 1)
q <- q / sum(q)

## See BF decrease
nvec <- c(stats::rmultinom(n = 1, size = 100, prob = q))
gl <- simgl(nvec = nvec)
rmbayesgl(gl = gl)

nvec <- c(stats::rmultinom(n = 1, size = 1000, prob = q))
gl <- simgl(nvec = nvec)
rmbayesgl(gl = gl)

}

}
\author{
David Gerard
}
