% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{simgl}
\alias{simgl}
\title{Stupid simulator for genotype log-likelihoods}
\usage{
simgl(nvec, sig = 0.1)
}
\arguments{
\item{nvec}{The genotype counts. \code{nvec[k]} contains the number of
individuals with genotype \code{k-1}.}

\item{sig}{The standard deviation. Higher means more uncertainty. You should
think of this as in units of dosages.}
}
\description{
Gets at the idea of genotype uncertainty, but don't use this for simulations.
I just created this to help me debug.
}
\author{
David Gerard
}
