% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theo.R
\name{theofreq}
\alias{theofreq}
\title{Theoretical frequencies at equilibrium.}
\usage{
theofreq(alpha, r, ploidy)
}
\arguments{
\item{alpha}{A numeric vector containing the double reduction parameter(s).
This should be a
vector of length \code{floor(ploidy/4)} where \code{alpha[i]}
is the probability of exactly \code{i} pairs of IBDR alleles
being in the gamete. Note that \code{sum(alpha)} should be less than
1, as \code{1 - sum(alpha)} is the probability of no double reduction.}

\item{r}{The allele frequency of the reference allele.}

\item{ploidy}{The ploidy of the species. This should be an even positive
integer.}
}
\value{
A list with the following elements
\describe{
\item{\code{p}}{The gamete frequencies at equilibrium.}
\item{\code{q}}{The genotype frequencies at equilibrium.}
}
}
\description{
These return gamete and genotype frequencies as calculated in
Huang et al (2019). Only supported for ploidies less than or equal to 10.
}
\references{
\itemize{
\item{Huang, K., Wang, T., Dunn, D. W., Zhang, P., Cao, X., Liu, R., & Li, B. (2019). Genotypic frequencies at equilibrium for polysomic inheritance under double-reduction. G3: Genes, Genomes, Genetics, 9(5), 1693-1706. \doi{10.1534/g3.119.400132}}
}
}
\author{
David Gerard
}
