% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustat.R
\name{uobj}
\alias{uobj}
\title{Objective used in \code{\link{hweustat}()}}
\usage{
uobj(nvec, alpha, omega = NULL)
}
\arguments{
\item{nvec}{A vector of length ploidy + 1. \code{nvec[i]} is the observed
number of individuals with dosage \code{i-1}.}

\item{alpha}{The double reduction parameter. Should be of length
\code{floor(ploidy/4)}. These should sum to less than 1, since
\code{1-sum(alpha)} is the assumed probability of no double
reduction.}

\item{omega}{A (ploidy+1) by (ploidy+1) positive semidefinite matrix.
This is an estimate of the \emph{inverse} covariance of the elements
of the U-statistic.}
}
\description{
Objective used in \code{\link{hweustat}()}
}
\author{
David Gerard
}
