% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe.R
\name{freqnext}
\alias{freqnext}
\title{Update genotype frequencies after one generation}
\usage{
freqnext(freq, alpha, segarray = NULL)
}
\arguments{
\item{freq}{The current genotype frequencies. This should be a
vector of length K+1, where K is the ploidy of the species.
\code{freq[i]} could contain the proportion of individuals
that have genotype \code{i-1}.}

\item{alpha}{The double reduction parameter(s). This should be a
vector of length \code{floor(ploidy/4)} where \code{alpha[i]}
is the probability of exactly \code{i} pairs of IBDR alleles
being in the gamete. Note that \code{sum(alpha)} should be less than
1, as \code{1 - sum(alpha)} is the probability of no double reduction.}

\item{segarray}{The output of \code{\link{zsegarray}()}. We will calculate
it if \code{segarray = NULL}. It is just an option so we don't
have to recalculate it very often if \code{alpha} is known.}
}
\value{
A vector of length \code{lenght(freq)} that contains the
updated genotype frequencies after one generation of random mating.
}
\description{
After one generation of random mating, update the genotype
frequencies.
}
\examples{
freq <- c(0.5, 0, 0, 0, 0.5)
freqnext(freq = freq, alpha = 0)

}
\author{
David Gerard
}
