% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{hwefit}
\alias{hwefit}
\title{Equilibrium and random mating estimation and testing for many loci.}
\usage{
hwefit(nmat, type = c("ustat", "mle", "rm", "nodr"), ...)
}
\arguments{
\item{nmat}{A matrix of counts. The rows index the loci and the columns
index the genotypes. So \code{nmat[i, j]} is the number of individuals
that have genotype \code{j-1} at locus \code{i}. The ploidy is
assumed to be \code{ncol(nmat)-1}.}

\item{type}{The method to use:
\describe{
\item{\code{"ustat"}}{U-statistic approach to test for equilibrium and
estimate double reduction rates givene equilibrium. The default.
See \code{\link{hweustat}()}.}
\item{\code{"mle"}}{Maximum likelihood estimation and testing. Only
supported for tetraploids. This will test for equilibrium and
random mating, and is generally better than \code{"ustat"}
for tetraploids. See \code{\link{hwetetra}()}.}
\item{\code{"rm"}}{Testing random mating, and estimating gamete
frequencies given random mating. See \code{\link{rmlike}()}.}
\item{\code{"nodr"}}{Testing equilibrium given no double reduction.
See \code{\link{hwenodr}()}.}
}}

\item{...}{Any other parameters to send to \code{\link{hweustat}()},
\code{\link{hwetetra}()},
\code{\link{rmlike}()}, or \code{\link{hwenodr}()}.}
}
\value{
A data frame. The columns of which can are described in
\code{\link{hwetetra}()}, \code{\link{hweustat}()},
\code{\link{rmlike}()}, or \code{\link{hwenodr}()}.
}
\description{
Estimates and tests for either equilibrium or random mating  across many loci
using \code{\link{hwetetra}()}, \code{\link{hweustat}()},
\code{\link{rmlike}()}, or \code{\link{hwenodr}()}.
}
\details{
We provide parallelization support through the \link[future]{future}
package.
}
\examples{
## Generate random data
set.seed(5)
ploidy <- 6
nloc <- 100
size <- 1000
r <- 0.5
alpha <- 1/12
qvec <- hwefreq(r = r, alpha = alpha, ploidy = ploidy)
nmat <- t(rmultinom(n = nloc, size = size, prob = qvec))

## Run the analysis in parallel on the local computer with two workers
future::plan(future::multisession, workers = 2)
hout <- hwefit(nmat = nmat,
               type = "ustat",
               thresh = 5)

## Shut down parallel workers
future::plan("sequential")

## Show that p-values are uniform
obs <- sort(hout$p_hwe)
plot(x = ppoints(n = length(obs)),
     y = obs,
     xlab = "theoretical",
     ylab = "observed",
     main = "qqplot")
abline(0, 1, col = 2, lty = 2)
mean(hout$p_hwe < 0.05, na.rm = TRUE)

## Consistent estimate for alpha
alpha
mean(hout$alpha1)

}
\author{
David Gerard
}
