% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{hwemom}
\alias{hwemom}
\title{Estimate double reduction parameter and test for HWE}
\usage{
hwemom(nvec, denom = c("expected", "observed"))
}
\arguments{
\item{nvec}{A vector containing the observed genotype counts,
where \code{nvec[[i]]} is the number of individuals with genotype
\code{i-1}. This should be of length \code{ploidy+1}.}

\item{denom}{What should we use for the denominator? The expected
genotype frequencies (\code{denom = "expected"}) or the
observed genotype frequencies (\code{denom = "observed"})?}
}
\value{
A list with the following elements
\describe{
\item{\code{alpha}}{The estimated double reduction parameter(s).
In diploids, this value is \code{NULL}.}
\item{\code{chisq_hwe}}{The chi-square test statistic for testing
against the null of equilibrium.}
\item{\code{df_hwe}}{The degrees of freedom associated with
\code{chisq_hwe}.}
\item{\code{p_hwe}}{The p-value against the null of equilibrium.}
\item{\code{chisq_alpha}}{The chi-square test statistic for
testing against the null of no double reduction (conditional
on equilibrium). In diploids, this value is \code{NULL}.}
\item{\code{df_alpha}}{The degrees of freedom associated with
\code{chisq_alpha}.}
\item{\code{p_alpha}}{The p-value against the null of no double
reduction (conditional on equilibrium). In diploids, this value
is \code{NULL}.}
}
}
\description{
Implements a generalized method-of-moments approach to test for
Hardy-Weinberg equilibrium in the presence of double reduction. This
also provides an estimate of double reduction given that the population
is at equilibrium.
}
\details{
For ploides greater than 4, not all gametic frequencies are possible
given rates of double reduction. That is, if the model for meiosis
from Huang et. al. (2019) is incorrect, then this is not an appropriate
test for HWE.
}
\examples{
## Diploid at various levels of deviation from HWE
hwemom(c(10, 20, 10))
hwemom(c(12, 16, 12))
hwemom(c(14, 12, 14))

## Tetraploid with strong deviation from HWE
nvec <- c(25, 0, 0, 0, 25)
hwemom(nvec)

## Hexaploid with exact frequencies at HWE
nvec <- round(hwefreq(r = 0.5,
                      alpha = 0.4,
                      ploidy = 6,
                      niter = 100,
                      tol = -Inf) * 100)
hwemom(nvec)

## Octoploid case with exact frequencies at HWE
nvec <- round(hwefreq(r = 0.5,
                      alpha = c(0.4, 0.1),
                      ploidy = 8,
                      niter = 100,
                      tol = -Inf) * 100)
hwemom(nvec)

}
\references{
\itemize{
\item{Huang, K., Wang, T., Dunn, D. W., Zhang, P., Cao, X., Liu, R., & Li, B. (2019). Genotypic frequencies at equilibrium for polysomic inheritance under double-reduction. G3: Genes, Genomes, Genetics, 9(5), 1693-1706.}
}
}
\author{
David Gerard
}
