% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwep-package.R
\docType{package}
\name{hwep-package}
\alias{hwep-package}
\alias{hwep}
\title{Hardy-Weinberg Equilibrium in Polyploids}
\description{
Inference concerning Hardy-Weinberg equilibrium (HWE) in
polyploids. Methods are available to test for HWE at any
ploidy level in the presence of double reduction. For
autopolyploid populations in HWE, methods are available
to estimate the degree of double reduction. We also
provide functions to calculate genotype frequencies at
equilibrium given rates of double reduction.
}
\section{Functions}{

The main functions are:
\describe{
\item{\code{\link{hwefit}()}}{Fit either \code{\link{hwetetra}()},
\code{\link{rmlike}()}, or \code{\link{hwemom}()} across
many loci. Parallelization is supported through the
\link{future} package.}
\item{\code{\link{hwetetra}()}}{Likelihood inference for tetraploids.
Includes tests for random mating, only small deviations from HWE,
and the presence of double reduction.}
\item{\code{\link{rmlike}()}}{Likelihood inference for random
mating in polyploids. This function tests for random mating
and estimates gametic frequencies given random mating.
This function does not assume a model for meiosis.
You should use \code{\link{hwetetra}()} for tetraploids.}
\item{\code{\link{hwemom}()}}{Generalized method of moments inference
for double reduction. This function tests for equilibrium given
double reduction rates and estimates these rates given equilibrium.
You should use \code{\link{hwetetra}()} for tetraploids.}
}

Other functions include:
\describe{
\item{\code{\link{dgamete}()}}{Gamete dosage probability given
parental dosage.}
\item{\code{\link{drbounds}()}}{Upper bounds on the rates of double
reduction given the complete equational segregation model.}
\item{\code{\link{freqnext}()}}{Update genotype frequencies after one
generation of random mating.}
\item{\code{\link{hwefreq}()}}{Generate Hardy-Weinberg equilibrium
genotype frequencies.}
\item{\code{\link{p_from_alpha}()}}{Obtain gamete frequencies from
the major allele frequency and double reduction rates.}
\item{\code{\link{zsegarray}()}}{All zygote dosage distributions
given all possible parental dosages.}
\item{\code{\link{zygdist}()}}{Zygote dosage distribution given
one pair of parental dosages.}
}
}

\section{Citation}{

If you find the methods in this package useful, please run the following
in R for citation information: \code{citation("hwep")}
}

\author{
David Gerard
}
